Final Export.sh Edits

+++ eval/export.sh
@@ -14,11 +14,74 @@ specified a guess is made and interactively verified.
 EOF
 }
 
+
+evaldir=$3
+graphdir=$4
+
 # selection of graphs
 sel=(
-	all-tiny/graphs/sched-by-count-cdf-all.eps
-	all-tiny/graphs/schedulability-cache-reuse-count-all.eps
-    )
+     	$graphdir/graphs/sched-by-count-cdf-all.eps
+	$graphdir/graphs/schedulability-cache-reuse-pct-all.eps
+	$graphdir/graphs/sched-by-count-cdf-approx.eps
+)
+
+
+if [ "$3" == "verify" ] 
+then
+	sel=(
+	     	prov-group-e/graphs/sched-by-count-cdf-all.eps
+		prov-group-e/graphs/schedulability-cache-reuse-pct-all.eps
+		prov-group-x/graphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+if [ "$3" == "simulation_small" ] 
+then
+	sel=(
+	     	all-small/graphs/sched-by-count-cdf-all.eps
+		all-small/graphs/schedulability-cache-reuse-pct-all.eps
+		all-small/graphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+if [ "$3" == "simulation_tiny" ] || [ "$3" == "quick" ]
+then
+	sel=(
+	     	all-tiny/graphs/sched-by-count-cdf-all.eps
+		all-tiny/graphs/schedulability-cache-reuse-pct-all.eps
+		all-tiny/graphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+if [ "$3" == "simulation_set_e" ] 
+then
+	sel=(
+	     	group-e/graphs/sched-by-count-cdf-all.eps
+		group-e/graphs/schedulability-cache-reuse-pct-all.eps
+		group-egraphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+if [ "$3" == "simulation_set_x" ] 
+then
+	sel=(
+	     	group-x/graphs/sched-by-count-cdf-all.eps
+		group-x/graphs/schedulability-cache-reuse-pct-all.eps
+		group-x/graphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+if [ "$3" == "all" ] 
+then
+	echo all
+	sel=(
+	     	group-e/graphs/sched-by-count-cdf-all.eps
+		group-e/graphs/schedulability-cache-reuse-pct-all.eps
+		group-x/graphs/sched-by-count-cdf-approx.eps
+	)
+fi
+
+
 
 ODIR=${1}
 DDIR=${2:-../../oleaf.git/}
@@ -46,7 +109,11 @@ if [ ! -d ${DDIR} ] ; then
 fi
 echo ${DDIR} exists
 
-DDIR=${DDIR}/graphs/
+
+
+DDIR=${DDIR}/"$3"/
+
+
 echo $DDIR
 mkdir -p ${DDIR}
 
@@ -58,7 +125,9 @@ for file in ${sel[@]} ; do
 	dst=${DDIR}/
 	dst+=$(basename $file)
 
-
-	echo "${src} -> ${dst}"
-	cp ${src} ${dst}
+	if [ -e ${src} ]
+	then
+		echo "${src} -> ${dst}"
+		cp ${src} ${dst}
+	fi
 done
